end ch8

+++ Program 8.2/Program 8.2/XYPoint.h
@@ -10,5 +10,5 @@
 
 @interface XYPoint : NSObject
 @property int x, y;
--(void) setX: (int) xVal andY: (int) yVal;
+-(void) setX: (float) xVal andY: (float) yVal;
 @end
