+++ Program 8.2/Program 8.2/XYPoint.m
@@ -11,7 +11,7 @@
 @implementation XYPoint
 @synthesize x, y;
 
--(void) setX: (int) xVal andY: (int) yVal
+-(void) setX: (float) xVal andY: (float) yVal
 {
     x = xVal;
     y = yVal;
